+++ testing/geckodriver/src/marionette.rs
@@ -208,7 +208,7 @@ impl MarionetteHandler {
             format!("Failed to set preferences: {}", e),
         ))?;
 
-        handler.prepare(&profile).map_err(|e| WebDriverError::new(
+        handler.prepare(&profile, options.env.unwrap_or_default()).map_err(|e| WebDriverError::new(
             ErrorStatus::UnknownError,
             e.to_string()
         ))?;
@@ -249,12 +249,6 @@ impl MarionetteHandler {
 
         let mut runner = FirefoxRunner::new(&binary, profile);
 
-        // https://developer.mozilla.org/docs/Environment_variables_affecting_crash_reporting
-        runner
-            .env("MOZ_CRASHREPORTER", "1")
-            .env("MOZ_CRASHREPORTER_NO_REPORT", "1")
-            .env("MOZ_CRASHREPORTER_SHUTDOWN", "1");
-
         // double-dashed flags are not accepted on Windows systems
         runner.arg("-marionette");
         if self.settings.jsdebugger {
@@ -263,6 +257,15 @@ impl MarionetteHandler {
         if let Some(args) = options.args.as_ref() {
             runner.args(args);
         }
+        if let Some(env) = options.env {
+            runner.envs(env);
+        }
+
+        // https://developer.mozilla.org/docs/Environment_variables_affecting_crash_reporting
+        runner
+            .env("MOZ_CRASHREPORTER", "1")
+            .env("MOZ_CRASHREPORTER_NO_REPORT", "1")
+            .env("MOZ_CRASHREPORTER_SHUTDOWN", "1");
 
         let browser_proc = runner.start().map_err(|e| {
             WebDriverError::new(